% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata-helpers.R
\name{attach_varlab}
\alias{attach_varlab}
\title{Append variable labels for Stata export}
\usage{
attach_varlab(tbl, labels, overwrite = TRUE)
}
\arguments{
\item{tbl}{A dataset}

\item{labels}{A set of levels. Either a tibble that has
two columns called `alias` and `description`, OR a vector
of the form `c(alias1 = "description1", alias2 = "description2)` that
can be coerced into one.}
}
\description{
Append variable labels for Stata export
}
\examples{
 mt_stata <- attach_varlab(
   as_tibble(mtcars),
   c(cyl = "Number of cylinders", hp = "Gross horsepower"))
 str(mt_stata)

vartab <- tribble(
 ~ cyl, ~hp,
 "Number of cylinders", "Gross horsepower"
)
 attach_varlab(
  as_tibble(mtcars),
  c(cyl = "Number of cylinders", hp = "Gross horsepower")
 )
}
