% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate_melt-cast.R
\name{cast_senate}
\alias{cast_senate}
\title{cast two two columns per case_id}
\usage{
cast_senate(tbl, cast_names)
}
\arguments{
\item{tbl}{A long dataset, where each row is a senator (senator-response)}

\item{cast_names}{The names of the variables in tbl that will be the \code{value.var} in dcast}

\item{names}{New names}
}
\description{
cast two two columns per case_id
}
