% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dataverse-ids.R
\docType{data}
\name{cces_dv_ids}
\alias{cces_dv_ids}
\title{Identifiers for datasets in CCES dataverse}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 5 columns.
}
\usage{
cces_dv_ids
}
\description{
Each rows corresponds to a dataverse datataset
}
\keyword{datasets}
