% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography-naming.R
\name{cd_from_acs}
\alias{cd_from_acs}
\title{get CD format from ACS}
\usage{
cd_from_acs(vec)
}
\arguments{
\item{vec}{a vector of strings from the ACS congressional district naming}
}
\description{
get CD format from ACS
}
\examples{
cd_from_acs("Congressional District 32 (115th Congress), California")

}
