% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-agreement.R
\name{code_threeway}
\alias{code_threeway}
\title{simpler version of dyad agreement, intended for long form where office is set}
\usage{
code_threeway(data, var1, var2)
}
\arguments{
\item{data}{dataset}

\item{var1}{first variable of Y, N, DK, or D,R}
}
\description{
simpler version of dyad agreement, intended for long form where office is set
}
\examples{
library(tibble)
df <- tibble(pid_self =   c("D", "R", "I", "D"),
             pid_actl =   c("R", "R", "R", "D"),
             issue_self = c("Y", "N", "DK", "N"),
             issue_actl = c("N", "N", "N", "Y"))

code_threeway(df, pid_self, pid_actl)
code_threeway(df, issue_self, issue_actl)


}
