% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-agreement.R
\name{dyad_agrmt}
\alias{dyad_agrmt}
\title{Compute dyadic agreement from dataframe.}
\usage{
dyad_agrmt(
  tbl,
  chamber,
  agrmt_name = "agrmt",
  svy_var = "response",
  policy_var = "vote"
)
}
\arguments{
\item{tbl}{dataframe}

\item{chamber}{One of \code{"H", "S", "Policy"}}

\item{agrmt_name}{The name for thew new variable in character}

\item{svy_var}{The opinion variable to look at}

\item{policy_var}{The vote variable to look at}
}
\description{
If vote was not held, assumes status quo. (i.e. doesn't drop)
}
