% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{H_to_house}
\alias{H_to_house}
\alias{H_to_cd}
\title{Recode H/S}
\usage{
H_to_house(x)

H_to_cd(x, h_to = "Congressional Districts", s_to = "States")
}
\description{
A convenience function to recode H to House (or CDs) and S to Senate
}
\examples{
library(tibble)
library(dplyr)
library(ggplot2)


tibble(chamber = c("H", "S")) |>
 ggplot() +
 geom_text(aes(label = chamber), x = 0.5, y = 0.5) +
 facet_wrap(~ chamber, labeller = labeller(chamber = H_to_house))

tibble(chamber = c("H", "S")) |>
 ggplot() +
 geom_text(aes(label = chamber), x = 0.5, y = 0.5) +
 facet_wrap(~ chamber, labeller = labeller(chamber = H_to_cd))

}
