% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatsvotes.R
\name{logodds}
\alias{logodds}
\title{Conversion of proportions to log of odds}
\usage{
logodds(x, clip = TRUE, clip_by = 0)
}
\arguments{
\item{x}{Numeric vector of proportions}

\item{clip}{Logical; if \code{TRUE}, clip values to avoid 0 and 1}

\item{clip_by}{Numeric offset from boundaries when clipping}
}
\value{
Log odds transformation of \code{x}
}
\description{
Conversion of proportions to log of odds
}
\keyword{internal}
