% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{my.parse}
\alias{my.parse}
\title{Parse no-space labels to cleaner words for graphing}
\usage{
my.parse(char, drop_year = TRUE)
}
\arguments{
\item{char}{text to parse}
}
\description{
Parse no-space labels to cleaner words for graphing
}
