% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-seats.R
\name{prop_seats}
\alias{prop_seats}
\title{Compute measure for seats relevant for seats-votes}
\usage{
prop_seats(tbl, pct_var = pct_yes2_raw, n_var = n_raw, group_var = q_label)
}
\description{
Compute measure for seats relevant for seats-votes
}
