% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_color_DR}
\alias{scale_color_DR}
\alias{scale_fill_DR}
\title{Red and Blue colors for D vs. R character}
\usage{
scale_color_DR(labels = c(R = "Rep.", D = "Dem."), reverse = FALSE, ...)

scale_fill_DR(labels = c(R = "Rep.", D = "Dem."), reverse = FALSE, ...)
}
\description{
Red and Blue colors for D vs. R character
}
\examples{
library(tibble)
library(ggplot2)

tibble(party = c("D", "R")) |>
 ggplot(aes(color = party)) +
 geom_text(aes(label = party), x = 0.5, y = 0.5, size = 12) +
 facet_wrap(~ party) +
 scale_color_DR()

}
\seealso{
ggredist::scale_color_party_d()
}
