% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-proportion-yes.R
\name{question_split}
\alias{question_split}
\alias{issue_split}
\title{Proportion Yes by question}
\usage{
question_split(tbl, ...)

issue_split(tbl, ...)
}
\arguments{
\item{tbl}{The dataset with the columns \code{cong}, \code{question}}

\item{...}{variables to group on in addition to year and question, tidyeval variables}

\item{tbl}{The dataset with the columns \code{cong}, \code{question}}

\item{...}{variables to group on in addition to year and question, tidyeval variables}
}
\description{
Splits by qID, or a question asked in a given year. Does not aggregate across years

Unlike \code{issue_split}, this question will aggregate across years if the same question is asked.
}
