% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{std_short_party}
\alias{std_short_party}
\title{Standardize party codes to single capital letters}
\usage{
std_short_party(x)
}
\arguments{
\item{x}{A vector of characters with unstandardized party, e.g. "Republican", "Not sure"..}
}
\value{
A vector of "D", "R",
}
\description{
Standardize party codes to single capital letters
}
\examples{
std_short_party(
 c("Republican",
   "Democrat",
   "Not sure",
   "Other Party / Independent",
   "Never Heard of Person")
   )


}
