% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stata.R
\name{vartab}
\alias{vartab}
\title{Recreate Stata's varaiable table}
\usage{
vartab(dta, string = NULL)
}
\arguments{
\item{dta}{output from \env{read_dta}}

\item{string}{string to search and filter, optional}
}
\description{
Recreate Stata's varaiable table
}
\examples{
# example data (what a read_dta tbl would look like)
df <- tibble(id = 1:3, pid = 1:3, age = 1:3)
attr(df[["id"]], "label") <- "Case Identifier"
attr(df[["pid"]], "label") <- "Partisan Identity"

vartab(df)
vartab(df, "identity")
}
