% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-agreement.R
\name{yes_no_agrmt}
\alias{yes_no_agrmt}
\title{simpler version of dyad agreement, intended for long form where office is set}
\usage{
yes_no_agrmt(var1, var2, data)
}
\arguments{
\item{var1}{first variable of Y, N, DK}

\item{data}{dataset}
}
\description{
simpler version of dyad agreement, intended for long form where office is set
}
