% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata-helpers.R
\name{attach_varlab}
\alias{attach_varlab}
\title{Append variable labels for Stata export}
\usage{
attach_varlab(tbl, labels, overwrite = TRUE)
}
\arguments{
\item{tbl}{A dataset}

\item{labels}{A set of levels. Either a tibble that has
two columns called `alias` and `description`, OR a vector
of the form `c(alias1 = "description1", alias2 = "description2)` that
can be coerced into one.}

\item{overwrite}{If \code{TRUE} (default), overwrite existing labels.}
}
\description{
Append variable labels for Stata export
}
\examples{
library(tibble)

 mt_stata <- as_tibble(mtcars) |>
  attach_varlab(c(cyl = "Number of cylinders", hp = "Gross horsepower"))

 # bulk edits by a tibble format
 vartab <- tribble(
  ~ alias, ~description,
  "cyl",  "Number of cylinders",
  "hp",  "Gross horsepower"
 )
 mt_stata <- as_tibble(mtcars) |>
  attach_varlab(vartab)


}
