% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_for-acs.R
\name{ccc_std_demographics}
\alias{ccc_std_demographics}
\title{Recode CCES variable values so that they merge to ACS variables}
\usage{
ccc_std_demographics(tbl)
}
\arguments{
\item{tbl}{A subset of the cumulative common content. Must include variables
\code{age}, \code{race}, \code{educ}, and \code{gender}.}
}
\description{
Recode CCES variable values so that they merge to ACS variables
}
