% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cmfmtW}
\alias{cmfmtW}
\title{comma formatting to char}
\usage{
cmfmtW(x, file, round = NA, pp = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{file}{path to write to}

\item{round}{round by}

\item{pp}{if true then multiply by one hundred}
}
\description{
comma formatting to char
}
