% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cmfmtW}
\alias{cmfmtW}
\title{comma formatting to char}
\usage{
cmfmtW(x, file, round = NA, pp = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{file}{path to write to}

\item{round}{round by}

\item{pp}{if true then multiply by one hundred}
}
\description{
comma formatting to char
}
\examples{
# Write a formatted count to a .tex file for LaTeX inclusion
cmfmtW(123456, tempfile(fileext = ".tex"))

# Write a rounded percentage
cmfmtW(0.6712, tempfile(fileext = ".tex"), round = 1, pp = TRUE)
}
