% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-agreement.R
\name{code_threeway}
\alias{code_threeway}
\title{Code threeway agreement (-1, 0, 1)}
\usage{
code_threeway(data, var1, var2)
}
\arguments{
\item{data}{dataset}

\item{var1}{first variable of Y, N, DK, or D, R}

\item{var2}{second variable of Y, N, DK, or D, R}
}
\description{
Simpler version of dyad agreement, intended for long form where office is set.
Returns 1 for agreement, -1 for disagreement, 0 for don't know / independent,
and NA if either input is missing.
}
\examples{
library(tibble)
df <- tibble(pid_self =   c("D", "R", "I", "D"),
             pid_actl =   c("R", "R", "R", "D"),
             issue_self = c("Y", "N", "DK", "N"),
             issue_actl = c("N", "N", "N", "Y"))

code_threeway(df, pid_self, pid_actl)
code_threeway(df, issue_self, issue_actl)


}
