% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{H_to_house}
\alias{H_to_house}
\alias{H_to_cd}
\alias{H_to_rep}
\title{Recode H/S}
\usage{
H_to_house(x)

H_to_cd(x, h_to = "Congressional Districts", s_to = "States")

H_to_rep(x, h_to = "House Representative", s_to = "Senator")
}
\arguments{
\item{x}{A character vector of chamber codes (\code{"H"}, \code{"S"})}

\item{h_to}{Label for House}

\item{s_to}{Label for Senate}
}
\description{
A convenience function to recode H to House (or CDs) and S to Senate
}
\examples{
H_to_house(c("H", "S"))
H_to_cd(c("H", "S"))
H_to_rep(c("H", "S"))

}
