% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt-cces.R
\name{melt_cces}
\alias{melt_cces}
\title{melt wide cces (raw versions) to correct var names}
\usage{
melt_cces(tbl, q_expr)
}
\arguments{
\item{tbl}{CCES dataset}

\item{q_expr}{regex that selects questions}
}
\description{
melt wide cces (raw versions) to correct var names
}
