% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcces-package.R
\docType{package}
\name{rcces-package}
\alias{rcces}
\alias{rcces-package}
\title{rcces: Tools to Melt, Combine, Clean, and Standardize CCES Questions}
\description{
Utility functions for analyzing and reshaping CCES dta files. This package is not meant for production. For better documented standard functions see `kuriwaki/ccesMRPprep`, `kuriwaki/ccesMRPrun`, and `kuriwaki/cces_cumulative`.
}
\author{
\strong{Maintainer}: Shiro Kuriwaki \email{shirokuriwaki@gmail.com}

}
\keyword{internal}
