% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcces-package.R
\docType{package}
\name{rcces-package}
\alias{rcces-package}
\title{Functions for CCES analysis}
\description{
Functions to melt, combine, clean, and standardize CCES questions
}
\author{
person("Shiro", "Kuriwaki", email = "kuriwaki@g.harvard.edu", role = c("aut", "cre"))
}
\keyword{package}
