% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcces-package.R
\docType{package}
\name{rcces-package}
\alias{rcces}
\alias{rcces-package}
\title{rcces: Helper Functions for CCES Representation Projects}
\description{
Functions to format, clean, and operationalize variables used in Shiro Kuriwaki's CCES-based representation research. Includes tools for reshaping survey responses, computing dyadic agreement between constituents and legislators, aggregating opinion to geographic units, labeling Stata variables, and writing formatted numbers for LaTeX output.
}
\author{
\strong{Maintainer}: Shiro Kuriwaki \email{shirokuriwaki@gmail.com}

}
\keyword{internal}
