% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-proportion-yes.R
\name{do_split}
\alias{do_split}
\alias{issue_split}
\alias{question_split}
\title{Computation underlying ques_split and issue_split}
\usage{
do_split(grptbl)

issue_split(tbl, ...)

question_split(tbl, ...)
}
\arguments{
\item{tbl}{The dataset with the columns \code{cong}, \code{question}}

\item{...}{variables to group on in addition to year and question, tidyeval variables}
}
\description{
Unlike \code{issue_split}, this question will aggregate across years if the same question is asked.

Splits by qID, or a question asked in a given year. Does not aggregate across years
}
