% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata-helpers.R
\name{vartab}
\alias{vartab}
\title{Recreate Stata's variable table}
\usage{
vartab(dta, string = NULL, name = alias)
}
\arguments{
\item{dta}{output from \env{read_dta} or \env{read_sav}}

\item{string}{string to search and filter, optional}

\item{name}{Name of the variable for the column of original stata variables}
}
\description{
Recreate Stata's variable table
}
\examples{
# example data (what a read_dta tbl would look like)
library(tibble)
df <- tibble(id = 1:3, pid = 1:3, age = 1:3)
attr(df[["id"]], "label") <- "Case Identifier"
attr(df[["pid"]], "label") <- "Partisan Identity"

vartab(df)
vartab(df, "identity")
}
